+++ paymentsheet-example/src/main/java/com/stripe/android/paymentsheet/example/samples/ui/customer/CustomerSheetViewState.kt
@@ -1,7 +1,7 @@
 package com.stripe.android.paymentsheet.example.samples.ui.customer
 
+import com.stripe.android.customersheet.CustomerEphemeralKey
 import com.stripe.android.customersheet.ExperimentalCustomerSheetApi
-import com.stripe.android.customersheet.repositories.CustomerEphemeralKey
 
 @OptIn(ExperimentalCustomerSheetApi::class)
 sealed class CustomerSheetViewState {
